A little clean-up
