+++ src/main/java/edu/harvard/h2ms/validator/EventValidator.java
@@ -77,8 +77,5 @@ public class EventValidator implements Validator {
 				errors.reject("Event.RequiredAnswer");
 			}
 		}
-	
-
 	}
-
 }
